+++ trunk/FE_DealII/dealii_fe_object.h
@@ -22,7 +22,7 @@
 #include <deal.II/lac/sparse_matrix.h>
 //#include <deal.II/lac/sparse_matrix.templates.h>
 //#include <deal.II/lac/solver_cg.h>
-#include <deal.II/lac/precondition.h>
+//#include <deal.II/lac/precondition.h>
 #include <deal.II/lac/constraint_matrix.h>
 #include <deal.II/grid/tria.h>
 #include <deal.II/grid/grid_generator.h>